Moved project from root context to workspace layer.
editor: Removed Document ProjectFile two way connection.
editor: Added Document.path
workspace: Added fileSystem.openDocumentInEditor
workspace: Updated Runnable support for QmlProgram.
application: Added Livekeys.main
application: Removed Livekeys.project
application: Moved QmlEngineInterceptor to ViewEngine.
base: Added Program and ProgramHolder.
base: Added Utf8.isEmpty.
editor: Added Document.isOnDisk
editor: Added Project.createDocument
editor: Updated Project.openFile interface.
editor: Added Project.monitorFiles
view: Added customizable interceptor (ViewEngineInterceptor).
view: Added QmlProgram.
view: Added QmlProjectInfo.

+++ lib/lveditor/src/linecontrol.cpp
@@ -488,7 +488,7 @@ void LineControl::applyOffsetToSectionsAfterIndex(unsigned sectionIndex, int off
         while (!q.empty())
         {
             LineSection* lsp = q.front(); q.pop();
-            for (int x = 0; x < lsp->children.size(); ++x)
+            for (size_t x = 0; x < lsp->children.size(); ++x)
                 q.push(&lsp->children[x]);
 
             if (offsetVisible) lsp->displayLineNumber += offset;
@@ -542,7 +542,7 @@ void LineControl::codeRemovalHandler(int pos, int removed)
                 while (true)
                 {
                     // finding which nested sections contain the left and right borders of the removed text
-                    for (int i = 0; i < parent->children.size(); ++i)
+                    for (size_t i = 0; i < parent->children.size(); ++i)
                     {
                         if (pos >= parent->children[i].startPosition && pos < parent->children[i].endPosition)
                             leftIndex = i;
@@ -741,7 +741,7 @@ void LineControl::codeAddingHandler(int pos, int added)
     while (!q.empty())
     {
         LineSection* lsp = q.front(); q.pop();
-        for (int x = 0; x < lsp->children.size(); ++x)
+        for (size_t x = 0; x < lsp->children.size(); ++x)
             q.push(&lsp->children[x]);
 
         // If the section contains the text addition point, we shift the round bound only
@@ -779,7 +779,7 @@ bool LineControl::handleOffsetsWithinASection(int sectionIndex, int delta)
     while (!q.empty())
     {
         LineSection* lsp = q.front(); q.pop();
-        for (int x = 0; x < lsp->children.size(); ++x)
+        for (size_t x = 0; x < lsp->children.size(); ++x)
             q.push(&lsp->children[x]);
 
         if (lsp->lineNumber <= m_firstDirtyLine && m_firstDirtyLine < lsp->lineNumber + lsp->lineSpan)
